requirements.txt
pandas
numpy
matplotlib
seaborn
sklearn
pyod
PYyaml